+++ src/layouts/InterfaceLayout/containers/InteractWithContractContainer/InteractWithContractContainer.vue
@@ -19,10 +19,11 @@
         </div>
         <div class="the-form domain-name">
           <input
+            v-ens-resolver="address"
             v-model="address"
             type="text"
-            name=""
             placeholder="Enter Domain Name or Address" >
+
           <i
             :class="[validAddress && address !== ''? '': 'not-good' ,'fa fa-check-circle good-button']"
             aria-hidden="true"
@@ -268,7 +269,8 @@ export default {
       raw: {},
       nonce: 0,
       signedTx: '',
-      transactionFee: 0
+      transactionFee: 0,
+      resolvedAddress: ''
     };
   },
   computed: {
@@ -282,16 +284,13 @@ export default {
     },
     selectedMethod(newVal) {
       this.writeInputs = {};
-      for (let inputs in newVal) {
+      for (const inputs in newVal) {
         if (newVal.hasOwnProperty(inputs)) {
           newVal.inputs.forEach(input => {
             this.$set(this.writeInputs, input.name, '');
           });
         }
       }
-    },
-    address(newVal) {
-      this.validAddress = this.$store.state.web3.utils.isAddress(newVal);
     },
     writeInputs: {
       // Watches nested values instead
@@ -333,9 +332,8 @@ export default {
         type.includes('bytes')
       ) {
         return 'string';
-      } else {
-        return 'number';
       }
+      return 'number';
     },
     selectFunction(method) {
       const contract = new this.$store.state.web3.eth.Contract(
@@ -428,7 +426,12 @@ export default {
           nonce: this.nonce,
           gasPrice: Number(unit.toWei(this.$store.state.gasPrice, 'gwei')),
           value: this.value,
-          to: this.address,
+          to:
+            this.resolvedAddress !== ''
+              ? this.resolvedAddress
+              : this.address !== ''
+                ? this.address
+                : '',
           data: this.data
         };
 
@@ -465,13 +468,12 @@ export default {
     },
     checkInputsFilled() {
       const inputs = Object.keys(this.writeInputs);
-      for (var i = 0; i < inputs.length; i++) {
+      for (let i = 0; i < inputs.length; i++) {
         if (this.writeInputs[inputs[i]] === '') {
           this.inputsFilled = false;
           return;
-        } else {
-          this.inputsFilled = true;
         }
+        this.inputsFilled = true;
       }
     }
   }