+++ src/router/index.js
@@ -4,12 +4,12 @@ import HomeLayout from '@/layouts/HomeLayout';
 import CreateWalletLayout from '@/layouts/CreateWalletLayout';
 import TeamLayout from '@/layouts/TeamLayout';
 import PrivacyPolicyLayout from '@/layouts/PrivacyPolicyLayout';
-import TermsOfConditionsLayout from '@/layouts/TermsOfConditionsLayout';
+import TermsAndConditionsLayout from '@/layouts/TermsAndConditionsLayout';
 import AccessWalletLayout from '@/layouts/AccessWalletLayout';
 import InterfaceLayout from '@/layouts/InterfaceLayout';
 import HelpCenterLayout from '@/layouts/HelpCenterLayout';
 import NotFoundLayout from '@/layouts/NotFoundLayout';
-import CreateWalletWarningLayout from '@/layouts/CreateWalletWarningLayout';
+import GettingStarted from '@/layouts/GettingStarted';
 import { router as routerConfig } from '@/configs/build';
 
 Vue.use(Router);
@@ -38,9 +38,9 @@ export default new Router({
       component: PrivacyPolicyLayout
     },
     {
-      path: '/terms-of-conditions',
-      name: 'TermsOfConditionsLayout',
-      component: TermsOfConditionsLayout
+      path: '/terms-and-conditions',
+      name: 'TermsAndConditionsLayout',
+      component: TermsAndConditionsLayout
     },
     {
       path: '/access-my-wallet',
@@ -59,8 +59,8 @@ export default new Router({
     },
     {
       path: '/getting-started',
-      name: 'CreateWalletWarningLayout',
-      component: CreateWalletWarningLayout
+      name: 'GettingStarted',
+      component: GettingStarted
     },
     {
       path: '*',
@@ -73,8 +73,7 @@ export default new Router({
       return {
         selector: to.hash
       };
-    } else {
-      window.scrollTo(0, 0);
     }
+    window.scrollTo(0, 0);
   }
 });