+++ src/translations/home/gb.js
@@ -32,7 +32,6 @@ const home = {
   faqsPrivKeyandPass: 'Private Keys And Passwords',
   socialTitle: 'Get Involved',
   socialSubheading: 'Your Cryptocurrency Masters Are Getting Hilariously Rich.',
-  promoTitle1: 'Your Security',
-  promoTitle2: 'Ethereum Wallet.'
+  promoTitle: "Ethereum's Original Wallet"
 };
 export default home;