+++ src/wallets/hardware/secalot/secalotWallet.js
@@ -16,7 +16,7 @@ export default class SecalotWallet extends HardwareWalletInterface {
     this.wallet = null;
     this.transport = null;
 
-    let options = opts || {};
+    const options = opts || {};
     this.addressToWalletMap = {};
     this.addressesToIndexMap = {};
     this.walletsRetrieved = [];
@@ -75,26 +75,23 @@ export default class SecalotWallet extends HardwareWalletInterface {
   getAddress() {
     if (this.wallet) {
       return this.wallet.address;
-    } else {
-      return null;
     }
+    return null;
   }
 
   getAddressString() {
     if (this.wallet) {
       return ethUtil.toChecksumAddress(this.getAddress());
-    } else {
-      return null;
     }
+    return null;
   }
 
   getAccounts() {
-    let _this = this;
+    const _this = this;
     if (arguments.length > 1 && typeof arguments[2] === 'function') {
       return _this.getMultipleAccounts(arguments[0], arguments[1]);
-    } else {
-      return _this._getAccounts();
     }
+    return _this._getAccounts();
   }
 
   getMultipleAccounts(count, offset) {
@@ -107,8 +104,8 @@ export default class SecalotWallet extends HardwareWalletInterface {
   }
 
   signMessage(msgData) {
-    let thisMessage = msgData.data ? msgData.data : msgData;
-    let app = new SecalotEth(this.transport, '');
+    const thisMessage = msgData.data ? msgData.data : msgData;
+    const app = new SecalotEth(this.transport, '');
     return new Promise(resolve => {
       resolve(app.signMessage(this.path, thisMessage));
     });
@@ -139,8 +136,8 @@ export default class SecalotWallet extends HardwareWalletInterface {
     typeof secalotSecret === 'string' ? Number(secalotSecret) : secalotSecret;
     return new Promise(resolve => {
       this.transport = new SecalotUsb();
-      let app = new SecalotEth(this.transport, secalotSecret);
-      let path = this.path;
+      const app = new SecalotEth(this.transport, secalotSecret);
+      const path = this.path;
       app.getAddress(path, (result, error) => {
         resolve(this.secalotCallback(result, error));
       });
@@ -148,7 +145,7 @@ export default class SecalotWallet extends HardwareWalletInterface {
   }
 
   createWallet(priv, pub, path, hwType, hwTransport) {
-    let wallet = {};
+    const wallet = {};
     if (typeof priv !== 'undefined') {
       wallet.privKey = priv.length === 32 ? priv : Buffer.from(priv, 'hex');
     }
@@ -176,7 +173,7 @@ export default class SecalotWallet extends HardwareWalletInterface {
   setHDAddressesHWWallet(start, limit) {
     this.walletsRetrieved = [];
     for (let i = start; i < start + limit; i++) {
-      let derivedKey = this.hdk.derive('m/' + i);
+      const derivedKey = this.hdk.derive('m/' + i);
       const tempWallet = this.createWallet(
         undefined,
         derivedKey.publicKey,
@@ -214,7 +211,7 @@ export default class SecalotWallet extends HardwareWalletInterface {
   // (Start) Internal methods underlying wallet usage methods
   async _getAccounts(count, offset) {
     return new Promise(resolve => {
-      let collect = {};
+      const collect = {};
       if (
         this.addressesToIndexMap[offset] &&
         this.addressesToIndexMap[offset + count - 1]
@@ -254,7 +251,7 @@ export default class SecalotWallet extends HardwareWalletInterface {
 
   signTxSecalot(rawTx) {
     return new Promise((resolve, reject) => {
-      let localCallback = (result, error) => {
+      const localCallback = (result, error) => {
         if (typeof error !== 'undefined') {
           error = error.errorCode ? u2f.getErrorByCode(error.errorCode) : error;
           reject(error);
@@ -263,14 +260,14 @@ export default class SecalotWallet extends HardwareWalletInterface {
         rawTx.v = this.sanitizeHex(result['v']);
         rawTx.r = this.sanitizeHex(result['r']);
         rawTx.s = this.sanitizeHex(result['s']);
-        let eTx_ = new EthereumjsTx(rawTx);
+        const eTx_ = new EthereumjsTx(rawTx);
         rawTx.rawTx = JSON.stringify(rawTx);
         rawTx.signedTx = this.sanitizeHex(eTx_.serialize().toString('hex'));
         rawTx.isError = false;
         resolve(rawTx);
       };
       // uiFuncs.notifier.info("Touch the LED for 3 seconds to sign the transaction. Or tap the LED to cancel.");
-      let app = new SecalotEth(this.transport, '');
+      const app = new SecalotEth(this.transport, '');
       const tx = new EthereumjsTx(rawTx);
       app.signTransaction(this.path, tx, localCallback);
     });
@@ -285,21 +282,17 @@ export default class SecalotWallet extends HardwareWalletInterface {
   // (End) Internal methods underlying wallet usage methods
   // (Start) Internal utility methods
   getNakedAddress(address) {
-    let naked = address.toLowerCase().replace('0x', '');
+    const naked = address.toLowerCase().replace('0x', '');
     if (naked.length % 2 === 0) {
       return naked.toString();
-    } else {
-      return '0' + naked.toString();
     }
+    return '0' + naked.toString();
   }
 
   _getAddressForWallet(wallet) {
     if (typeof wallet.pubKey === 'undefined') {
       return '0x' + ethUtil.privateToAddress(wallet.privKey).toString('hex');
-    } else {
-      return (
-        '0x' + ethUtil.publicToAddress(wallet.pubKey, true).toString('hex')
-      );
     }
+    return '0x' + ethUtil.publicToAddress(wallet.pubKey, true).toString('hex');
   }
 }