+++ src/wallets/web3-overide/web3-wallet-adapter.js
@@ -16,33 +16,29 @@ export default class Web3WalletAdapter {
   getPrivateKey() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPrivateKey();
-    } else {
-      return null;
     }
+    return null;
   }
 
   getPrivateKeyString() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPrivateKeyString();
-    } else {
-      return null;
     }
+    return null;
   }
 
   getPublicKey() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPublicKey();
-    } else {
-      return null;
     }
+    return null;
   }
 
   getPublicKeyString() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPublicKeyString();
-    } else {
-      return null;
     }
+    return null;
   }
 
   getAddress() {
@@ -52,11 +48,10 @@ export default class Web3WalletAdapter {
   getAddressString() {
     const address = this.wallet.getAddress();
     if (typeof address !== 'string') {
-      let rawAddress = '0x' + address.toString('hex');
+      const rawAddress = '0x' + address.toString('hex');
       return ethUtil.toChecksumAddress(rawAddress);
-    } else {
-      return address;
     }
+    return address;
   }
 
   getChecksumAddressString() {
@@ -68,33 +63,29 @@ export default class Web3WalletAdapter {
   get privateKey() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPrivateKeyString();
-    } else {
-      return null;
     }
+    return null;
   }
 
   get publicKey() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPublicKeyString();
-    } else {
-      return null;
     }
+    return null;
   }
 
   get privateKeyBuffer() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPrivateKey();
-    } else {
-      return null;
     }
+    return null;
   }
 
   get publicKeyBuffer() {
     if (this.privateKeyAvailable()) {
       return this.wallet.getPublicKey();
-    } else {
-      return null;
     }
+    return null;
   }
 
   get accounts() {