+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Feature
+- Implement ENS resolver directive, move directives to its own folder [#56](https://github.com/MyEtherWallet/MyEtherWallet/pull/56)
 - Intercept and override signTransaction and signMessage to force confirm modal display and approval [#51](https://github.com/MyEtherWallet/MyEtherWallet/pull/51)
 - Integrate Ledger hardware wallet [#48](https://github.com/MyEtherWallet/MyEtherWallet/pull/48)
 - Feature/verify message [#52](https://github.com/MyEtherWallet/MyEtherWallet/pull/52)
@@ -9,6 +10,8 @@
 - Add Create Wallet warning page [#46](https://github.com/MyEtherWallet/MyEtherWallet/pull/46)
 
 ### Devop
+- Add stale integration config [#70](https://github.com/MyEtherWallet/MyEtherWallet/pull/70)
+- Refactor network modal and sidemenu [#64](https://github.com/MyEtherWallet/MyEtherWallet/pull/64)
 - Cleanup linting error [#54](https://github.com/MyEtherWallet/MyEtherWallet/pull/54)
 - Merge all open PRS [#53](https://github.com/MyEtherWallet/MyEtherWallet/pull/53)
 - Add Network logo and change dot colors [#49](https://github.com/MyEtherWallet/MyEtherWallet/pull/49)